device refactor: move execute and apply into ForestDevice
