+++ pennylane_forest/qvm.py
@@ -96,82 +96,6 @@ class QVMDevice(ForestDevice):
         elif isinstance(device, str):
             return get_qc(device, as_qvm=True, noisy=noisy, **kwargs)
 
-    def execute(self, circuit, **kwargs):
-        if self.parametric_compilation:
-            self._circuit_hash = circuit.graph.hash
-
-        return super().execute(circuit, **kwargs)
-
-    def apply(self, operations, **kwargs):
-        """Run the QVM"""
-        # pylint: disable=attribute-defined-outside-init
-        if self.parametric_compilation and "pyqvm" not in self.qc.name:
-            self.apply_parametric_program(operations, **kwargs)
-        else:
-            super().apply(operations, **kwargs)
-
-        prag = Program(Pragma("INITIAL_REWIRING", ['"PARTIAL"']))
-
-        if self.active_reset:
-            prag += RESET()
-
-        self.prog = prag + self.prog
-
-        qubits = sorted(self.wiring.values())
-        ro = self.prog.declare("ro", "BIT", len(qubits))
-        for i, q in enumerate(qubits):
-            self.prog.inst(MEASURE(q, ro[i]))
-
-        self.prog.wrap_in_numshots_loop(self.shots)
-
-    def apply_parametric_program(self, operations, **kwargs):
-        """Applies a parametric program by applying parametric
-        operation with symbolic parameters.
-        """
-        # pylint: disable=attribute-defined-outside-init
-        rotations = kwargs.get("rotations", [])
-
-        # Storing the active wires
-        self._active_wires = ForestDevice.active_wires(operations + rotations)
-
-        # Apply the circuit operations
-        for i, operation in enumerate(operations):
-            # map the operation wires to the physical device qubits
-            device_wires = self.map_wires(operation.wires)
-
-            if i > 0 and operation.name in ("QubitStateVector", "BasisState"):
-                raise DeviceError(
-                    "Operation {} cannot be used after other Operations have already been applied "
-                    "on a {} device.".format(operation.name, self.short_name)
-                )
-
-            # Prepare for parametric compilation
-            par = []
-            for param in operation.data:
-                if getattr(param, "requires_grad", False) and operation.name != "BasisState":
-                    # Using the idx for trainable parameter objects to specify the
-                    # corresponding symbolic parameter
-                    parameter_string = "theta" + str(id(param))
-
-                    if parameter_string not in self._parameter_reference_map:
-                        # Create a new PyQuil memory reference and store it in the
-                        # parameter reference map if it was not done so already
-                        current_ref = self.prog.declare(parameter_string, "REAL")
-                        self._parameter_reference_map[parameter_string] = current_ref
-
-                    # Store the numeric value bound to the symbolic parameter
-                    self._parameter_map[parameter_string] = [param.unwrap()]
-
-                    # Appending the parameter reference to the parameters
-                    # of the corresponding operation
-                    par.append(self._parameter_reference_map[parameter_string])
-                else:
-                    par.append(param)
-
-            self.prog += self._operation_map[operation.name](*par, *device_wires.labels)
-
-        self.prog += self.apply_rotations(rotations)
-
     def generate_samples(self):
         if self.parametric_compilation:
             for region, value in self._parameter_map.items():
